# Phase2/tfm/Swagger/openapi-tspa.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "XFSC-TRAIN",
    "description": "This is the REST API of the XFSC TRAIN TSPA.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:16003",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/tspa/v1/{framework-name}/trust-list/tsp": {
      "put": {
        "tags": [
          "trust-list-publication-controller"
        ],
        "operationId": "createTSP",
        "parameters": [
          {
            "name": "framework-name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TSP"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful operation. Creates a TSP record in the trust list.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters supplied or Bad Request"
          },
          "401": {
            "description": "UnAuthorized Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/tspa/v1/init/xml/{framework-name}/trust-list": {
      "put": {
        "tags": [
          "trust-list-publication-controller"
        ],
        "operationId": "createTrustListXML",
        "parameters": [
          {
            "name": "framework-name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/JSONTrustListInit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "UnAuthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/tspa/v1/init/json/{framework-name}/trust-list": {
      "put": {
        "tags": [
          "trust-list-publication-controller"
        ],
        "operationId": "createTrustListJSON",
        "parameters": [
          {
            "name": "framework-name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JSONTrustListInit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "UnAuthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/tspa/v1/{framework-name}/trust-list/tsp/{id}": {
      "delete": {
        "tags": [
          "trust-list-publication-controller"
        ],
        "operationId": "deleteTSP",
        "parameters": [
          {
            "in": "path",
            "name": "framework-name",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation. Deletes a TSP record in the trust list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "UnAuthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "trust-list-publication-controller"
        ],
        "operationId": "updateTSP",
        "parameters": [
          {
            "in": "path",
            "name": "framework-name",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "description": "update a tsp in trust list within a given trust framework",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TSP"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation. Updates a TSP record in the trust list. Returns the patched trust-list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "UnAuthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/tspa/v1/{framework-name}/vc/trust-list": {
      "get": {
        "tags": [
          "trust-list-publication-controller"
        ],
        "operationId": "getVC",
        "parameters": [
          {
            "name": "framework-name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifiableCredentialDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/tspa/v1/{framework-name}/trust-list": {
      "get": {
        "tags": [
          "trust-list-publication-controller"
        ],
        "operationId": "getTrustList",
        "parameters": [
          {
            "in": "path",
            "name": "framework-name",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrustServiceStatusList"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/TrustServiceStatusList"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "trust-list-publication-controller"
        ],
        "operationId": "deleteTrustList",
        "parameters": [
          {
            "in": "path",
            "name": "framework-name",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation."
          },
          "400": {
            "description": "Invalid parameters supplied or Bad Request"
          },
          "401": {
            "description": "Unauthorized to delete trust list from given framework"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "response": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "JSONTrustListInit": {
        "type": "object",
        "properties": {
          "TrustServiceStatusList": {
            "type": "object",
            "properties": {
              "SchemeInformation": {
                "type": "object",
                "properties": {
                  "TSLVersionIdentifier": {
                    "type": "string"
                  },
                  "TSLSequenceNumber": {
                    "type": "string"
                  },
                  "TSLType": {
                    "type": "string"
                  },
                  "SchemeOperatorName": {
                    "type": "object",
                    "properties": {
                      "Name": {
                        "type": "string"
                      }
                    }
                  },
                  "SchemeOperatorAddress": {
                    "type": "object",
                    "properties": {
                      "PostalAddresses": {
                        "type": "object",
                        "properties": {
                          "PostalAddress": {
                            "type": "object",
                            "properties": {
                              "StreetAddress": {
                                "type": "string"
                              },
                              "Locality": {
                                "type": "string"
                              },
                              "PostalCode": {
                                "type": "string"
                              },
                              "CountryName": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      },
                      "ElectronicAddress": {
                        "type": "object",
                        "properties": {
                          "URI": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "SchemeName": {
                    "type": "object",
                    "properties": {
                      "Name": {
                        "type": "string"
                      }
                    }
                  },
                  "SchemeInformationURI": {
                    "type": "object",
                    "properties": {
                      "URI": {
                        "type": "string"
                      }
                    }
                  },
                  "SchemeTypeCommunityRules": {
                    "type": "object",
                    "properties": {
                      "URI": {
                        "type": "string"
                      }
                    }
                  },
                  "SchemeTerritory": {
                    "type": "string"
                  },
                  "PolicyOrLegalNotice": {
                    "type": "object",
                    "properties": {
                      "TSLLegalNotice": {
                        "type": "string"
                      }
                    }
                  },
                  "PointersToOtherTSL": {
                    "type": "string"
                  },
                  "ListIssueDateTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "NextUpdate": {
                    "type": "object",
                    "properties": {
                      "dateTime": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "TSP": {
        "type": "object",
        "properties": {
          "TrustServiceProvider": {
            "type": "object",
            "properties": {
              "TSPCurrentStatus": {
                "type": "string"
              },
              "StatusStartingTime": {
                "type": "string"
              },
              "TSPInformation": {
                "type": "object",
                "properties": {
                  "TSPName": {
                    "type": "string"
                  },
                  "TSPLegalName": {
                    "type": "string"
                  },
                  "TSPRole": {
                    "type": "string"
                  },
                  "TrustSchemeName": {
                    "type": "string"
                  },
                  "OtherTSL": {
                    "type": "string"
                  },
                  "TSPInformationURI": {
                    "type": "string"
                  },
                  "TSPLegalBasis": {
                    "type": "string"
                  },
                  "TSPEntityIdentifierList": {
                    "type": "object",
                    "properties": {
                      "TSPEntityIdendifier": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "Type": {
                              "type": "string"
                            },
                            "Value": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  },
                  "TSPCertificationList": {
                    "type": "object",
                    "properties": {
                      "TSPCertification": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "Type": {
                              "type": "string"
                            },
                            "Value": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  },
                  "TSPKeywords": {
                    "type": "string"
                  },
                  "Address": {
                    "type": "object",
                    "properties": {
                      "ElectronicAddress": {
                        "type": "string"
                      },
                      "PostalAddress": {
                        "type": "object",
                        "properties": {
                          "StreetAddress1": {
                            "type": "string"
                          },
                          "StreetAddress2": {
                            "type": "string"
                          },
                          "City": {
                            "type": "string"
                          },
                          "State": {
                            "type": "string"
                          },
                          "Country": {
                            "type": "string"
                          },
                          "PostalCode": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              },
              "SubmitterInfo": {
                "type": "object",
                "properties": {
                  "Name": {
                    "type": "string"
                  },
                  "Address": {
                    "type": "object",
                    "properties": {
                      "ElectronicAddress": {
                        "type": "string"
                      },
                      "PostalAddress": {
                        "type": "object",
                        "properties": {
                          "StreetAddress1": {
                            "type": "string"
                          },
                          "StreetAddress2": {
                            "type": "string"
                          },
                          "City": {
                            "type": "string"
                          },
                          "State": {
                            "type": "string"
                          },
                          "Country": {
                            "type": "string"
                          },
                          "PostalCode": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              },
              "TSPServices": {
                "type": "object",
                "properties": {
                  "TSPService": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "ServiceCurrentStatus": {
                          "type": "string"
                        },
                        "StatusStartingTime": {
                          "type": "string"
                        },
                        "ServiceName": {
                          "type": "string"
                        },
                        "ServiceTypeIdentifier": {
                          "type": "string"
                        },
                        "ServiceSupplyPoint": {
                          "type": "string"
                        },
                        "ServiceDefinitionURI": {
                          "type": "string"
                        },
                        "ServiceDigitalIdentity": {
                          "type": "object",
                          "properties": {
                            "Value": {
                              "type": "string"
                            },
                            "KeyType": {
                              "type": "string"
                            }
                          }
                        },
                        "AdditionalServiceInformation": {
                          "type": "object",
                          "properties": {
                            "ServiceIssuedCredentialTypes": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "CredentialType": {
                                    "type": "string"
                                  }
                                }
                              }
                            },
                            "ServiceGovernanceURI": {
                              "type": "string"
                            },
                            "ServiceBusinessRulesURI": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}